Add files via upload
To develop an ontology that effectively captures our meta-model shown in section 4.3, Basic Formal Ontology
(BFO is selected as our foundational ontology because it provides a robust metaphysical framework that naturally
aligns with the distinct types of entities in models. This Ontology can be developed by our algorithm to improve
interoperability, reusability, and clarity of the conceptual relationships.
Initially, we have merged our ontology with BFO, however, it was found that this file includes all BFO classes
that may not be necessarily needed in our case. Consequently, we aim to extract only the relevant classes that we were
required for our ontology. By doing this, a slim and streamlined version is developed that contains only essential BFO
classes. This gives us a lightweight ontology that contains only the required files.